# docs/api/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CRM System API
  description: |-
    API completa para o CRM System, incluindo gestão de clientes, tarefas, atividades, email, helpdesk e integração com WhatsApp.
    
    Esta API fornece acesso programático a todas as funcionalidades do sistema, seguindo os princípios RESTful com autenticação JWT.
  version: 1.0.0
  contact:
    name: CRM System Support
    email: api-support@crm-system.com
    url: https://crm-system.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.crm-system.com/v1
    description: Production server
  - url: https://staging-api.crm-system.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /helpdesk/departments:
    get:
      tags:
        - Helpdesk
      summary: Listar Departamentos
      description: Retorna uma lista de todos os departamentos
      operationId: listDepartments
      responses:
        '200':
          description: Lista de departamentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Helpdesk
      summary: Criar Departamento
      description: Cria um novo departamento no helpdesk
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Departamento criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /helpdesk/tickets:
    get:
      tags:
        - Helpdesk
      summary: Listar Tickets
      description: Retorna uma lista paginada de tickets do helpdesk
      operationId: listTickets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: status
          in: query
          description: Filtro por status do ticket
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: department
          in: query
          description: Filtro por departamento
          schema:
            type: string
        - name: priority
          in: query
          description: Filtro por prioridade
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        '200':
          description: Lista de tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HelpdeskTicket'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Helpdesk
      summary: Criar Ticket
      description: Cria um novo ticket no helpdesk
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/HelpdeskTicket'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'

  # WhatsApp
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Número da página
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Itens por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      description: Termo de busca
      schema:
        type: string
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: string
          example: "123"
        name:
          type: string
          example: "TI - Suporte Técnico"
        description:
          type: string
          example: "Suporte técnico e infraestrutura"
        isRecursive:
          type: boolean
          example: true
        groups:
          type: array
          items:
            type: string
          example: ["1", "2", "3"]
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDepartmentRequest:
      type: object
      required:
        - name
        - description
        - groups
      properties:
        name:
          type: string
          example: "TI - Suporte Técnico"
        description:
          type: string
          example: "Suporte técnico e infraestrutura"
        isRecursive:
          type: boolean
          default: false
        groups:
          type: array
          items:
            type: string
          example: ["1", "2", "3"]
    HelpdeskTicket:
      type: object
      properties:
        id:
          type: string
          example: "123"
        title:
          type: string
          example: "Problema com acesso ao sistema"
        description:
          type: string
          example: "Usuário não consegue fazer login no sistema"
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
          example: "open"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        department:
          type: string
          example: "TI - Suporte Técnico"
        contactName:
          type: string
          example: "João Silva"
        contactPhone:
          type: string
          example: "+5511999999999"
        assigneeId:
          type: string
          example: "456"
        tags:
          type: array
          items:
            type: string
          example: ["login", "acesso"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTicketRequest:
      type: object
      required:
        - title
        - description
        - department
      properties:
        title:
          type: string
          example: "Problema com acesso ao sistema"
        description:
          type: string
          example: "Usuário não consegue fazer login no sistema"
        department:
          type: string
          example: "TI - Suporte Técnico"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: "medium"
        contactName:
          type: string
          example: "João Silva"
        contactPhone:
          type: string
          example: "+5511999999999"
        assigneeId:
          type: string
          example: "456"
        tags:
          type: array
          items:
            type: string
          example: ["login", "acesso"]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Descrição do erro"
        details:
          type: string
          example: "Detalhes adicionais do erro"
        code:
          type: string
          example: "ERROR_CODE"
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...